Reorganized changes to spi-master
